Restructure project into packages
