+++ src/ui/main.py
@@ -1,6 +1,6 @@
 import sys
 from PyQt6.QtWidgets import QApplication
-from ui_mainwindow import ABTestWindow
+from .ui_mainwindow import ABTestWindow
 
 if __name__ == "__main__":
     app = QApplication(sys.argv)
